pyproject.toml
[project]
name = "helicon"
version = "0.4.0"
description = "GPU-Accelerated Magnetic Nozzle Simulation & Detachment Analysis Toolkit"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "h5py>=3.8",
]

[project.scripts]
helicon = "helicon.cli:main"

[project.optional-dependencies]
mlx = ["mlx>=0.10"]
optimize = ["scikit-learn>=1.3"]
botorch = ["botorch>=0.9", "torch>=2.0"]
dev = ["pytest>=7.0", "scikit-learn>=1.3"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 95

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E741"]  # allow I for current (physics convention)

[tool.ty.environment]
python-version = "3.11"

[tool.ty.rules]
unresolved-import = "ignore"

[dependency-groups]
dev = [
    "scikit-learn>=1.8.0",
]
